Calculate Baseline Nutrient Inadequacy (AFE Method)

This function calculates the baseline inadequacy of nutrients for different administrative groups using the Adequate Food Energy (AFE) Method.

calculateBaselineInadequacyAfe(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  nctListDf = nctList,
  intakeThresholdsDf = intakeThresholds,
  aggregationGroup = c("admin0Name", "admin1Name"),
  MNList = c("Ca", "Carbohydrates", "Cu", "Energy", "Fat", "Fe", "Fibre", "I", "IP6",
    "Mg", "Protein", "Se", "Zn", "Ash", "B6", "B2", "D", "N", "K", "P", "Moisture",
    "Cholesterol", "E", "Na", "A", "C", "B12", "B1", "B3", "B9", "B5", "B7", "Mn")
)

Arguments

householdConsumptionDf

A dataframe containing household consumption data. Must contain columns: "householdId", "amountConsumedInG", "memberCount".

householdDetailsDf

A dataframe containing household details. Must contain column: "householdId".

nctListDf

A dataframe containing nutrient composition tables. Must contain columns: "nutrient", "foodId".

intakeThresholdsDf

A dataframe containing intake thresholds for nutrients. Must contain columns: "nutrient", "CND".

aggregationGroup

A character vector of administrative groups to aggregate the data. Must not be empty. Defaults to c("admin0Name", "admin1Name").

MNList

A character vector of nutrients to be included in the analysis. If empty, defaults to a comprehensive list of nutrients.

Value

A dataframe with the baseline inadequacy of nutrients for the specified administrative groups.

Examples

calculateBaselineInadequacyAfe(
    householdConsumptionDf = householdConsumption,
    householdDetailsDf = householdDetails,
    nctListDf = nctList,
    intakeThresholdsDf = intakeThresholds,
    aggregationGroup = c("admin0Name", "admin1Name"),
    MNList = c("Ca", "Carbohydrates")
)
#> Error in eval(expr, envir, enclos): object 'householdConsumption' not found