Calculate Pre and Post Fortification Summaries

This function calculates summaries of nutrient inadequacy before and after different fortification methods for different administrative groups using either the Adult Female Equivalent (AFE) Method, the Critical Nutrient Density (CND) Method, or both.

Usage

calculate_pre_and_post_fortification_summaries(
  householdConsumptionDf = householdConsumption,
  householdDetailsDf = householdDetails,
  nctListDf = nctList,
  intakeThresholdsDf = intakeThresholds,
  aggregationGroup = c("admin0Name", "admin1Name"),
  fortifiableFoodItemsDf = fortifiableFoodItem,
  foodVehicleName = "wheat flour",
  fortificationLevelsDf = fortificationLevels,
  years = c(2021:2024),
  MNList = "A",
  metric = "afe",
  method = "LSFF"
)

Arguments

householdConsumptionDf

A dataframe containing household consumption data. Must contain columns: "householdId", "amountConsumedInG".

householdDetailsDf

A dataframe containing household details. Must contain columns: "householdId", "memberCount".

nctListDf

A dataframe containing nutrient composition tables. Must contain column: "nutrient".

intakeThresholdsDf

A dataframe containing intake thresholds for nutrients. Must contain columns: "nutrient", "CND", "CUL".

aggregationGroup

A character vector of administrative groups to aggregate the data. Must not be empty. Defaults to c("admin0Name", "admin1Name").

fortifiableFoodItemsDf

A dataframe containing fortifiable food items.`.

foodVehicleName

A character string specifying the name of the food vehicle for fortification. Defaults to "wheat flour".

fortificationLevelsDf

A dataframe containing the fortification levels for different nutrients and years.

years

A numeric vector specifying the years for which LSFF is analyzed. Defaults to 2021:2024.

MNList

A character vector of nutrients to be included in the analysis. Defaults to "A". Must not be empty.

metric

A character string specifying the method to be used for the analysis. Must be one of "AFE", "CND", or "BOTH". Defaults to "AFE".

method

A character string specifying the fortification method to be used for the analysis. Must be one of "LSFF", "BIO", or "FERT". Defaults to "LSFF".

Value

A dataframe containing summaries of nutrient inadequacy before and after fortification for different administrative groups.

Examples

if (FALSE) { # \dontrun{
calculate_pre_and_post_fortification_summaries(
    householdConsumptionDf = householdConsumption,
    householdDetailsDf = householdDetails,
    nctListDf = nctList,
    intakeThresholdsDf = intakeThresholds,
    aggregationGroup = c("admin0Name", "admin1Name"),
    fortifiableFoodItemsDf = createFortifiableFoodItemsTable(),
    foodVehicleName = "wheat flour",
    fortificationLevelsDf = fortificationLevels,
    years = c(2021:2024),
    MNList = c("A", "Ca"),
    metric = "AFE",
    method = "LSFF"
)
} # }