Year Average Fortification Level

This function loads the average fortification levels from a CSV file located in the data directory, and then returns the average fortification level for a specific year and micronutrient.

Usage

yearAverageFortificationLevel(
  fortification_vehicle,
  Year,
  MN,
  fortificationLevels
)

Arguments

fortification_vehicle

A string specifying the fortification vehicle.

Year

An integer specifying the year.

MN

A string specifying the micronutrient.

fortificationLevels

A dataframe containing the average fortification levels.

Value

A numeric value representing the average fortification level for the specified year and micronutrient.

Examples

if (FALSE) { # \dontrun{
avgFortificationLevel <- yearAverageFortificationLevel("wheat flour", 2024, "A")
} # }